+++ bin/miden-cli/src/lib.rs
@@ -9,6 +9,7 @@ use miden_client::account::AccountHeader;
 use miden_client::auth::TransactionAuthenticator;
 use miden_client::builder::ClientBuilder;
 use miden_client::keystore::FilesystemKeyStore;
+use miden_client::note_transport::grpc::GrpcNoteTransportClient;
 use miden_client::store::{NoteFilter as ClientNoteFilter, OutputNoteRecord};
 use miden_client::{Client, DebugMode, IdPrefixFetchError};
 use miden_client_sqlite_store::SqliteStore;
@@ -187,9 +188,17 @@ impl Cli {
             builder = builder.max_block_number_delta(delta);
         }
 
-        let mut client = builder.build().await?;
+        if let Some(tl_config) = cli_config.note_transport {
+            let client = GrpcNoteTransportClient::connect(
+                tl_config.endpoint.to_string(),
+                tl_config.timeout_ms,
+            )
+            .await
+            .map_err(|e| CliError::Client(e.into()))?;
+            builder = builder.note_transport(Arc::new(client));
+        }
 
-        client.ensure_genesis_in_place().await?;
+        let client = builder.build().await?;
 
         // Execute CLI command
         match &self.action {