+++ crates/web-client/test/playwright.global.setup.ts
@@ -97,6 +97,7 @@ export const test = base.extend<{ forEachTest: void }>({
             TransactionScript,
             TransactionScriptInputPair,
             TransactionScriptInputPairArray,
+            TransactionSummary,
             Word,
             WebClient,
             MockWebClient,
@@ -178,6 +179,7 @@ export const test = base.extend<{ forEachTest: void }>({
           window.TransactionScriptInputPair = TransactionScriptInputPair;
           window.TransactionScriptInputPairArray =
             TransactionScriptInputPairArray;
+          window.TransactionSummary = TransactionSummary;
           window.WebClient = WebClient;
           window.Word = Word;
           window.MockWebClient = MockWebClient;
@@ -188,6 +190,7 @@ export const test = base.extend<{ forEachTest: void }>({
 
           // Add the remote prover url to window
           window.remoteProverUrl = proverUrl;
+          window.rpcUrl = rpcUrl;
           if (window.remoteProverUrl) {
             window.remoteProverInstance =
               window.TransactionProver.newRemoteProver(window.remoteProverUrl);